# swagger.yml
swagger: "2.0"
info:
  description: This is the swagger file that goes with our server code
  version: "1.0.0"
  title: Swagger REST Article
consumes:
  - "application/json"
produces:
  - "application/json"

basePath: "/"

# Paths supported by the server application
paths:
  /product:
    post:
      operationId: "app.routes.product.add_product"
      tags:
        - "Product"
      summary: "A new product add by a client"
      description: "Add a new product"
      parameters:
        - in: "body"
          name: "body"
          description: "Product object that needs to be added to the database"
          required: true
          schema:
            $ref: "#/definitions/Product"
      responses:
        200:
          description: "Successful add a new product operation"
          schema:
            $ref: "#/definitions/Product"
        400:
          description: "Duplicate data entry"
        500:
          description: "Internal Server Error"

    get:
      operationId: "app.routes.product.get_products"
      tags:
        - "Product"
      summary: "Products added by clients"
      description: "Read the list of products"
      responses:
        200:
          description: "Successful read product list operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Product"

  /product/{productId}:
    get:
      operationId: "app.routes.product.get_product"
      tags:
        - "Product"
      summary: "A product added by a client"
      description: "Read the product with id"
      parameters:
        - name: "productId"
          in: "path"
          description: "ID of product to return"
          required: true
          type: "integer"
          format: "int64"
      responses:
        200:
          description: "Successful read product operation"
          schema:
            $ref: "#/definitions/Product"
    put:
      operationId: "app.routes.product.update_product"
      tags:
        - "Product"
      summary: "A product updated by a client"
      description: "Update the product with id"
      parameters:
        - name: "productId"
          in: "path"
          description: "ID of product to return"
          required: true
          type: "integer"
          format: "int64"
      responses:
        200:
          description: "Successful update product operation"
          schema:
            $ref: "#/definitions/Product"

        500:
          description: "Internal Server Error"

    delete:
      operationId: "app.routes.product.delete_product"
      tags:
        - "Product"
      summary: "A product delete by a client"
      description: "Delete the product with id"
      parameters:
        - name: "productId"
          in: "path"
          description: "ID of product to return"
          required: true
          type: "integer"
          format: "int64"
      responses:
        200:
          description: "Successful delete product operation"
          schema:
            $ref: "#/definitions/Product"

        500:
          description: "Internal Server Error"

definitions:
  Product:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      description:
        type: "string"
      price:
        type: "number"
        format: "float"
      quantity:
        type: "integer"
        format: "int32"